refactor: Rename customize_your_product_screen directory
